Version specificly for 2023 Update Revit compatibility and refactor codebase
- Added `GetElementIdValue` method in `CollisionEngine.cs` to handle different Revit versions.
- Updated category ID retrieval in `Connect.cs` to use `IntegerValue`.
- Modified `AllowElement` method in `Connect.cs` for consistency.
- Changed output path and dependencies in `RibbonSetup.csproj` for Revit 2023.
- Updated scale value parsing in `SheetCreate.cs` to use `int.Parse`.
- Refactored collision handling in `SleevePlacer.cs`, including method name changes for consistency.
- Improved clarity in `RoundUpToNearestIncrement` method in `SleevePlacer.cs`.
- Added null checks for family objects in `SleeveSettings.cs`.
- Updated `packages.config` to use Revit 2023 API version.

+++ Sheet Creator/SheetCreate.cs
@@ -295,7 +295,7 @@ namespace Intech
                 try
                 {
                     //change viewport to correct scale
-                    viewport.ChangeTypeId(new ElementId(Int64.Parse(scales[sheetscale.Remove(0, 1).Remove(sheetscale.Length - 2, 1)].Item1)));
+                    viewport.ChangeTypeId(new ElementId(int.Parse(scales[sheetscale.Remove(0, 1).Remove(sheetscale.Length - 2, 1)].Item1)));
                 }
                 catch
                 {
